Also fix fetch_size type mismatch

+++ sdks/python/apache_beam/io/jdbc.py
@@ -120,7 +120,7 @@ Config = typing.NamedTuple(
      ('connection_init_sqls', typing.Optional[typing.List[str]]),
      ('read_query', typing.Optional[str]),
      ('write_statement', typing.Optional[str]),
-     ('fetch_size', typing.Optional[int]),
+     ('fetch_size', typing.Optional[np.int16]),
      ('output_parallelization', typing.Optional[bool]),
      ('autosharding', typing.Optional[bool]),
      ('partition_column', typing.Optional[str]),
